Corrected multiple development environment problem
Corrected problem described on the readme. 

The fix which was presented worked fine if you have only a windows enviromnent, however, if you develop on both windows and unix-like platforms this solution alone is just not enough. We need to identify the current platform the user is running the command to make the changes accordingly.